Moving failing / failed notifications to correct activity
